requirements.txt
astroid==2.4.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycrypto==2.6.1
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
